# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Neutron Star API",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
        "/calculator/ipv4": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "calculator"
                ],
                "summary": "解析 ipv4",
                "parameters": [
                    {
                        "type": "string",
                        "name": "ip_addr",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/calculator.IPv4"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/commonresp.Err"
                        }
                    }
                }
            }
        },
        "/calculator/ipv6": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "calculator"
                ],
                "summary": "解析 ipv6",
                "parameters": [
                    {
                        "type": "string",
                        "name": "ip_addr",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/calculator.IPv6"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/commonresp.Err"
                        }
                    }
                }
            }
        },
        "/pixiv": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "pixiv"
                ],
                "summary": "获取 pixiv 图片 url",
                "parameters": [
                    {
                        "type": "integer",
                        "name": "num",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/pixiv.Images"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/commonresp.Err"
                        }
                    }
                }
            }
        },
        "/user/login": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "登录",
                "parameters": [
                    {
                        "description": "body",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/userreq.User"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/userresp.Token"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/commonresp.Err"
                        }
                    }
                }
            }
        },
        "/user/me": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "个人信息",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/userresp.User"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/commonresp.Err"
                        }
                    }
                }
            }
        },
        "/user/signup": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "注册",
                "parameters": [
                    {
                        "description": "body",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/userreq.User"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/commonresp.Ok"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/commonresp.Err"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "calculator.IPv4": {
            "type": "object",
            "properties": {
                "cidr": {
                    "type": "string"
                },
                "count": {
                    "type": "integer"
                },
                "end": {
                    "type": "string"
                },
                "ip": {
                    "type": "string"
                },
                "mask_bits": {
                    "type": "integer"
                },
                "start": {
                    "type": "string"
                },
                "subnet_mask": {
                    "type": "string"
                },
                "wildcard_mask": {
                    "type": "string"
                }
            }
        },
        "calculator.IPv6": {
            "type": "object",
            "properties": {
                "long": {
                    "$ref": "#/definitions/calculator.SubIPv6"
                },
                "short": {
                    "$ref": "#/definitions/calculator.SubIPv6"
                }
            }
        },
        "calculator.SubIPv6": {
            "type": "object",
            "properties": {
                "cidr": {
                    "type": "string"
                },
                "count": {
                    "type": "integer"
                },
                "end": {
                    "type": "string"
                },
                "ip": {
                    "type": "string"
                },
                "mask_bits": {
                    "type": "integer"
                },
                "start": {
                    "type": "string"
                }
            }
        },
        "commonresp.Err": {
            "type": "object",
            "properties": {
                "detail": {
                    "type": "string"
                }
            }
        },
        "commonresp.Ok": {
            "type": "object"
        },
        "pixiv.Images": {
            "type": "object",
            "properties": {
                "img_urls": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "userreq.User": {
            "type": "object",
            "properties": {
                "password": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "userresp.Token": {
            "type": "object",
            "properties": {
                "token": {
                    "type": "string"
                }
            }
        },
        "userresp.User": {
            "type": "object",
            "properties": {
                "username": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "token",
            "in": "header"
        }
    }
}